+++ tools/ports/regal.py
@@ -110,6 +110,9 @@ def get(ports, settings, shared):
       '-Wno-unused-parameter',
       '-Wno-nontrivial-memaccess',
       '-fdelayed-template-parsing',
+      # src/boost/boost/print/string_list.hpp calls std::string(NULL), whose
+      # constructor is declared _Nonnull
+      '-Wno-nonnull',
     ]
     if settings.PTHREADS:
       flags += ['-pthread']