+++ ChangeLog.md
@@ -20,7 +20,7 @@ See docs/process.md for more on how version tagging works.
 
 4.0.24 (in development)
 -----------------------
-- compiler-rt, libcxx, and libcxxabi were updated to LLVM 21.1.8. (#26405,
+- compiler-rt, libcxx, and libcxxabi were updated to LLVM 21.1.8. (#26045,
   #26058)
 - A new `-sEXECUTABLE` setting was added which adds a #! line to the resulting
   JavaScript and makes it executable.  This setting defaults to true when the